세션 모집 공고 전체 수정

+++ src/main/java/com/umc/bscene/domain/session/service/impl/SessionRecruitmentCommandServiceImpl.java
@@ -42,6 +42,7 @@ public class SessionRecruitmentCommandServiceImpl implements SessionRecruitmentC
                 .orElseThrow(() -> new BandException(BandErrorCode.BAND_MEMBER_NOT_FOUND));
 
         Band band = bandMember.getBand();
+        validateBandOwner(band, userId);
         validateDeadlineAt(request.getDeadlineAt());
 
         SessionRecruitment recruitment = SessionRecruitment.builder()
@@ -74,13 +75,7 @@ public class SessionRecruitmentCommandServiceImpl implements SessionRecruitmentC
 
         Band band = recruitment.getBand();
 
-        bandMemberRepository
-                .findByBand_IdAndUser_IdAndStatus(
-                        band.getId(),
-                        userId,
-                        BandMemberStatus.ACCEPTED
-                )
-                .orElseThrow(() -> new SessionException(SessionErrorCode.BAND_PERMISSION_DENIED));
+        validateBandOwner(band, userId);
 
         validateDeadlineAt(request.getDeadlineAt());
 
@@ -110,13 +105,7 @@ public class SessionRecruitmentCommandServiceImpl implements SessionRecruitmentC
 
         Band band = recruitment.getBand();
 
-        bandMemberRepository
-                .findByBand_IdAndUser_IdAndStatus(
-                        band.getId(),
-                        userId,
-                        BandMemberStatus.ACCEPTED
-                )
-                .orElseThrow(() -> new SessionException(SessionErrorCode.BAND_PERMISSION_DENIED));
+        validateBandOwner(band, userId);
 
         sessionRecruitmentRepository.delete(recruitment);
     }
@@ -124,6 +113,12 @@ public class SessionRecruitmentCommandServiceImpl implements SessionRecruitmentC
     private void validateDeadlineAt(LocalDateTime deadlineAt) {
         if (!deadlineAt.isAfter(LocalDateTime.now())) {
             throw new SessionException(SessionErrorCode.INVALID_SESSION_RECRUITMENT_DEADLINE);
+        }
+    }
+
+    private void validateBandOwner(Band band, Long userId) {
+        if (!band.getOwner().getId().equals(userId)) {
+            throw new SessionException(SessionErrorCode.BAND_PERMISSION_DENIED);
         }
     }
 }
